Clamp npc settings, translate every screen, fix quiet loot loss, 1.29.9
Fifth review pass over the mod. Four defects and two hardening gaps.

Numbers on an npc's own settings were read straight out of the tag while the
boss configuration has been swept for that since 1.29.4. Two of them had a
bound that existed only in the editor: keepDistance is squared by the goal that
reads it, and an int square overflows past 46340 blocks into a radius of one or
a negative the check can never pass; transitionLengthTicks goes into GeckoLib's
blend length, where a value that never runs out is a model frozen halfway
between two poses. Both now clamp on the way in, as does a NaN yaw on a spawn
point or a totem, which a plain Mth.clamp lets through.

Every screen goes through the language files. Forty-seven labels, buttons and
picker headings were English literals, so both locale files were complete and
the editor was not. The picker resolves its heading before centring it, because
measuring the key rather than the sentence puts the heading off to one side.

Boss loot that no chest claimed disappeared in silence. The sweep stays - the
case it exists for is a death another mod cancelled, where the boss is still
standing with its inventory and spilling would hand out a second copy - but it
says what it dropped, and a boss that really died with its chest switched off
in between now drops it on the floor instead of waiting the timeout out.

Cocoons and their guards walked the whole level per boss, twice, next to the
index built to make that one walk per level per tick; they read the index now.
The capture, tether and cocoon managers are asked whether they hold anything
before they are ticked, like the ten schedulers beside them.

A mixin that no longer fits the installed CustomNPCs build was a warning, not a
failure: the addon loaded and did nothing, and the version range for that mod
is open on purpose. The config is required now, injectors default to require 1,
and a startup check reports which of the nine hooks are in place.

Tests: the range sweep is shared and now covers the npc settings classes with a
range written out per key and floats poisoned with NaN and both infinities, so
a number saved with no stated range fails rather than joining the unchecked
ones; no screen may hold display text instead of a key; the startup check must
name every interface the mixins add. 87 unit tests, 51 gametests, all green.

+++ src/main/java/com/goodbird/cnpcgeckoaddon/ai/KeepDistanceGoal.java
@@ -32,7 +32,9 @@ public class KeepDistanceGoal extends Goal {
             return false;
         }
         this.target = entity;
-        return npc.distanceToSqr(entity) < (double) (distance * distance);
+        // Squared as a double: an int square overflows past 46340 blocks and reads back as a
+        // radius of one, or as a negative one the check can never pass.
+        return npc.distanceToSqr(entity) < (double) distance * (double) distance;
     }
 
     @Override
